Pixel format handling updates and misc fixes/tweaks
* Imagine: Simplify supported window pixel formats to RGB565 and RGBA8888
* Imagine: Add PixelDesc::reversed() and nativeOrder()
* Imagine: Remove unneeded PIXEL_RGBX8888 and PIXEL_ABGR8888 formats
* Imagine: Store an explicit format in AndroidWindow and simplify setFormat()
* Imagine: Check if config supports EGL_GL_COLORSPACE_SRGB before setting it
* Imagine: Don't implicitly call resetAppData() and resetRendererData() in Window's onExit delegate and instead rely on proper destructor ordering
* Imagine: Detach all windows in Renderer destructor to correctly destroy all drawables that reference it
* Imagine: Add MultiChoiceMenuItem::updateDisplayString()
* Imagine: Remove unneeded java reflection backward compatibility in BaseActivity
* EmuFramework: Add option to set the render pixel format for EmuVideo
* EmuFramework: Add EmuSystem::renderFramebuffer() to output the current system's framebuffer without advancing emulation if supported
* EmuFramework: Rename EmuSystem::onPrepareVideo() to onRequestedVideoFormatChange() and call only on format changes
* EmuFramework: Add EmuVideo::startFrameWithAltFormat() for automatic conversion to the render pixel format
* EmuFramework: Clear EmuVideo's texture in EmuSystem::closeRuntimeSystem()
* EmuFramework: Check that the texture is initialized in EmuVideo::resetImage()
* EmuFramework: Disable SRGB color space option if the window format is RGB565
* EmuFramework: Add optional validator argument to readOptionValue()
* 2600.emu, C64.emu, GBA.emu, MD.emu, NES.emu, PCE.emu: Add support for native rendering in RGBA8888 format
* C64.emu: Delay calling vice_init() until onMainWindowCreated() so video_arch_canvas_init() has the proper bpp value
* MD.emu: Render to a fixed size frame buffer to optimize remap_line()

+++ MSX.emu/src/blueMSX/VideoChips/VDP.c
@@ -1201,7 +1201,7 @@ static void initPalette(VDP* vdp)
     }
 
     for (i = 0; i < 256; i++) {
-        vdp->paletteFixed[i] = videoGetColor(255 * ((i >> 2) & 7) / 7, 
+        vdp->paletteFixed[i] = videoGetColor(255 * ((i >> 2) & 7) / 7,
                                            255 * ((i >> 5) & 7) / 7, 
                                            255 * ((i & 3) == 3 ? 7 : 2 * (i & 3)) / 7);
     }
@@ -1517,7 +1517,7 @@ static void loadState(VDP* vdp)
 
     for (index = 0; index < sizeof(vdp->palette) / sizeof(vdp->palette[0]); index++) {
         sprintf(tag, "vdp->palette%d", index);
-        vdp->palette[index] = (UInt16)saveStateGet(state, tag, 0);
+        vdp->palette[index] = saveStateGet(state, tag, 0);
     }
 
     saveStateGetBuffer(state, "vram", vdp->vram, 0x4000 * vdp->vramPages);
